+++ config.cfg
@@ -1,6 +1,6 @@
-MAIL_SERVER = 'tigerpairmailserver'
-MAIL_USERNAME = 'ourusername'
-MAIL_PASSWORD = 'ourpassword'
-MAIL_PORT = ??? (I think we can choose this)
+MAIL_SERVER = 'smtp.gmail.com'
+MAIL_USERNAME = 'tigerpaircontact@gmail.com'
+MAIL_PASSWORD = 'wearecool'
+MAIL_PORT = 465
 MAIL_USE_SSL= True
 MAIL_USE_TLS = False
