+++ site/pages/alumni/index.html
@@ -22,40 +22,49 @@
 		<!-- Begin page content -->
 		<main role="main" class="flex-shrink-0">
 			<div class="container">
-				{% if matched: %}
-					<p>You have been matched with {{matchname}}!</p>
-					<p>Please reach out to them at <a href="{{matchemail}}">{{matchemail}}</a></p>
-				{% else: %}
-					<br/>
-					<h1>Alumni Info</h1>
-					<p>Update your info with the desired data and press Submit.</p>
-					<br/>
+				{% if exists == False: %}
+					<p>put your email lol</p>
 					<form method="POST" action="#">
-						<label for="firstname">First Name</label>
-						<input class="form-control mr-sm-2" type="text" id="firstname" name="firstname" value="{{firstname}}" required>
-						<br/>
-						<label for="lastname">Last Name</label>
-						<input class="form-control mr-sm-2" type="text" id="lastname" name="lastname" value="{{lastname}}" required>
-						<br/>
 						<label for="email">Email</label>
 						<input class="form-control mr-sm-2" type="text" id="email" name="email" value="{{email}}" required>
-						<br/>
-						<label for="major">Major</label>
-						<select class="form-control mr-sm-2" id="major" name="major" required>
-							<option value="{{major}}" disabled selected hidden>{{major}}</option>
-						</select>
-						<br/>
-						<label for="career">Career Field</label>
-						<!-- Changed id from field to career -->
-						<select class="form-control mr-sm-2" id="career" name="career" required>
-							<option value="{{career}}" disabled selected hidden>{{career}}</option>
-							<option value="accounting">Accounting</option>
-							<option value="advertising">Advertising</option>
-							<option value="other">Other</option>
-						</select>
-						<br/>
 						<button class="btn btn-outline-success my-2 my-sm-0" type="submit">Submit</button>
 					</form>
+				{% else: %}
+					{% if matched: %}
+						<p>You have been matched with {{matchname}}!</p>
+						<p>Please reach out to them at <a href="{{matchemail}}">{{matchemail}}</a></p>
+					{% else: %}
+						<br/>
+						<h1>Alumni Info</h1>
+						<p>Update your info with the desired data and press Submit.</p>
+						<br/>
+						<form method="POST" action="#">
+							<label for="firstname">First Name</label>
+							<input class="form-control mr-sm-2" type="text" id="firstname" name="firstname" value="{{firstname}}" required>
+							<br/>
+							<label for="lastname">Last Name</label>
+							<input class="form-control mr-sm-2" type="text" id="lastname" name="lastname" value="{{lastname}}" required>
+							<br/>
+							<label for="email">Email</label>
+							<input class="form-control mr-sm-2" type="text" id="email" name="email" value="{{email}}" required>
+							<br/>
+							<label for="major">Major</label>
+							<select class="form-control mr-sm-2" id="major" name="major" required>
+								<option value="{{major}}" disabled selected hidden>{{major}}</option>
+							</select>
+							<br/>
+							<label for="career">Career Field</label>
+							<!-- Changed id from field to career -->
+							<select class="form-control mr-sm-2" id="career" name="career" required>
+								<option value="{{career}}" disabled selected hidden>{{career}}</option>
+								<option value="accounting">Accounting</option>
+								<option value="advertising">Advertising</option>
+								<option value="other">Other</option>
+							</select>
+							<br/>
+							<button class="btn btn-outline-success my-2 my-sm-0" type="submit">Submit</button>
+						</form>
+					{% endif %}
 				{% endif %}
 			</div>
 		</main>
